pyproject.toml
[tool.poetry]
name = "skypro-hw-31"
version = "0.1.0"
description = ""
authors = ["Andrey Torlopov <mail@andreytorlopov.ru>"]
readme = "README.md"
packages = [{include = "skypro_hw_31"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
autopep8 = "^2.0.2"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
isort = "^5.12.0"
pyformat = "^0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'shmavito.settings'
python_files = [
    'tests.py',
    'test_*.py',
    '*_tests.py'
]